tutoriales-vision-u/functions/requirements.txt
flask
firebase-functions
firebase-admin
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
blinker